+++ src/stories/avatar/index.tsx
@@ -2,6 +2,7 @@ import { Avatar as ZendeskAvatar } from "@zendeskgarden/react-avatars";
 import { theme } from "../theme";
 import styled from "styled-components";
 import { AvatarArgs } from "./_types";
+// import useWindowSize from "../../hooks/useWindowSize";
 
 const UgAvatar = styled(ZendeskAvatar)`
   ${(props) => {
@@ -19,7 +20,27 @@ const UgAvatar = styled(ZendeskAvatar)`
  */
 const Avatar = (props: AvatarArgs) => {
   const fixedBadge = props.badge && props.badge > 9 ? "9+" : props.badge;
-  return <UgAvatar {...props} badge={fixedBadge} />;
+  const wrapChildren = (type: string) => {
+    if (type === "icon") return props.children;
+    if (type === "image")
+      return <img alt="avatar" src={props.children as string} />;
+    if (type === "text") return <Avatar.Text>{props.children}</Avatar.Text>;
+  };
+
+  // const { width } = useWindowSize();
+
+  return (
+    <UgAvatar
+      {...props}
+      badge={fixedBadge}
+      children={wrapChildren(props.avatarType || "text")}
+      size={props.size || "small"}
+      // size={
+      //   props.size ??
+      //   (width && width > parseInt(theme.breakpoints.sm) ? "small" : "extrasmall")
+      // }
+    />
+  );
 };
 Avatar.Text = UgAvatar.Text;