+++ src/stories/buttons/split-button/index.stories.tsx
@@ -6,7 +6,7 @@ import { Menu } from "../../dropdowns/menu";
 import { Trigger } from "../../trigger";
 import { Button } from "../button";
 import { IconButton } from "../icon-button";
-import { ReactComponent as LeafIcon } from "@zendeskgarden/svg-icons/src/16/chevron-down-stroke.svg";
+import { ReactComponent as LeafIcon } from "../../../assets/icons/chevron-down-stroke.svg";
 import { SplitButton } from ".";
 import { SplitButtonArgs } from "./_types";
 
@@ -51,4 +51,8 @@ Default.parameters = {
 export default {
   title: "Atoms/Buttons/SplitButton",
   component: SplitButton,
+  parameters: {
+    // Sets a delay for the component's stories
+    chromatic: { delay: 300 },
+  },
 } as ComponentMeta<typeof IconButton>;