+++ src/stories/counter/_types.tsx
@@ -1,0 +1,8 @@
+import { TagArgs } from "../tags/_types";
+
+
+
+export interface CounterArgs extends TagArgs {
+  status: "progress" | "completed" | "incoming" | "functional" | "experiential";
+  counter: number;
+}