+++ src/stories/logo/_types.tsx
@@ -2,15 +2,9 @@ import { ImgHTMLAttributes } from 'react';
 
 export interface LogoArgs extends ImgHTMLAttributes<HTMLImageElement> {
   /** Specifies the kind of logo */
-  type: Type;
+  type: 'vertical' | 'horizontal' | 'icon';
   /** Specifies the css class */
-  className: string;
+  className?: string;
   /** Sets the image dimension */
-  size: number;
-}
-
-export enum Type {
-  vertical = 'vertical',
-  horizontal = 'horizontal',
-  icon = 'icon'
+  size?: number;
 }