+++ src/stories/navigation/content/index.tsx
@@ -1,0 +1,9 @@
+import { Content as ZendeskContent } from "@zendeskgarden/react-chrome";
+import { ContentArgs } from "./_types";
+
+/**
+ * A Content defines the main content of an HTML document which displays on the browser
+ */
+const Content = (props: ContentArgs) => <ZendeskContent {...props} />;
+
+export { Content };