+++ src/stories/typography/code/index.stories.tsx
@@ -45,4 +45,8 @@ export default {
       control: "text",
     },
   },
+  parameters: {
+    // Sets a delay for the component's stories
+    chromatic: { delay: 300 },
+  },
 } as ComponentMeta<typeof Code>;